Add files via upload
5/5 sin norminette

+++ push_swap.h
@@ -6,7 +6,7 @@
 /*   By: crmunoz- <crmunoz-@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/05/30 17:08:59 by crmunoz-          #+#    #+#             */
-/*   Updated: 2024/07/24 18:12:30 by crmunoz-         ###   ########.fr       */
+/*   Updated: 2024/07/24 20:36:39 by crmunoz-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -33,7 +33,7 @@ typedef struct s_stack
 	struct s_stack	*next;
 }					t_stack;
 
-int		ft_atol(char *str);
+long		ft_atol(char *str);
 char	**ft_split(char const *s, char c);
 size_t	ft_strlen(const char *s);
 t_stack	*ft_lstlast(t_stack *lst);
